Add authorization&authentication with JWT
Read Auth headers and add token to the context
Check in resolvers context for user info to allow/deny requests
Add Role field to the User
Send register email with the link to confirm account

+++ src/resolvers/follow.js
@@ -10,6 +10,9 @@ const followResolver = {
       context,
       info
     ) => {
+      if (!context.user) {
+        throw new Error('Unathenticated');
+      }
       try {
         // Need to add check if the follower is the authorized user and this follow is connected to him
         // Check if already following?
@@ -34,6 +37,9 @@ const followResolver = {
       }
     },
     deleteFollow: async (parent, { input: { id } }) => {
+      if (!context.user) {
+        throw new Error('Unathenticated');
+      }
       try {
         // Need to add check if the follower is the authorized user and this follow was created by him
         const follow = await Follow.findByIdAndRemove(id);
